refactor: #1497 Rename from Allophone to Sound
- Rename `/content` controllers
