+++ src/main/java/ai/elimu/util/csv/CsvContentExtractionHelper.java
@@ -17,7 +17,7 @@ import ai.elimu.model.v2.enums.content.allophone.SoundType;
 import ai.elimu.model.v2.gson.content.StoryBookChapterGson;
 import ai.elimu.model.v2.gson.content.StoryBookGson;
 import ai.elimu.model.v2.gson.content.StoryBookParagraphGson;
-import ai.elimu.web.content.sound.AllophoneCsvExportController;
+import ai.elimu.web.content.sound.SoundCsvExportController;
 import ai.elimu.web.content.emoji.EmojiCsvExportController;
 import ai.elimu.web.content.letter.LetterCsvExportController;
 import ai.elimu.web.content.letter_sound_correspondence.LetterSoundCorrespondenceCsvExportController;
@@ -49,7 +49,7 @@ public class CsvContentExtractionHelper {
     private static final Logger logger = LogManager.getLogger();
     
     /**
-     * For information on how the CSV files were generated, see {@link AllophoneCsvExportController#handleRequest}.
+     * For information on how the CSV files were generated, see {@link SoundCsvExportController#handleRequest}.
      */
     public static List<Allophone> getAllophonesFromCsvBackup(File csvFile) {
         logger.info("getAllophonesFromCsvBackup");
