+++ src/main/java/ai/elimu/web/content/sound/SoundEditController.java
@@ -14,32 +14,34 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
-@RequestMapping("/content/sound/create")
-public class AllophoneCreateController {
+@RequestMapping("/content/sound/edit")
+public class SoundEditController {
     
     private final Logger logger = LogManager.getLogger();
     
     @Autowired
     private AllophoneDao allophoneDao;
 
-    @RequestMapping(method = RequestMethod.GET)
-    public String handleRequest(Model model) {
+    @RequestMapping(value = "/{id}", method = RequestMethod.GET)
+    public String handleRequest(Model model, @PathVariable Long id) {
     	logger.info("handleRequest");
         
-        Allophone allophone = new Allophone();
+        Allophone allophone = allophoneDao.read(id);
         model.addAttribute("allophone", allophone);
         
         model.addAttribute("soundTypes", SoundType.values());
 
-        return "content/sound/create";
+        return "content/sound/edit";
     }
     
-    @RequestMapping(method = RequestMethod.POST)
+    @RequestMapping(value = "/{id}", method = RequestMethod.POST)
     public String handleSubmit(
+            @PathVariable Long id,
             @Valid Allophone allophone,
             BindingResult result,
             Model model,
@@ -49,14 +51,14 @@ public class AllophoneCreateController {
         
         if (StringUtils.isNotBlank(allophone.getValueIpa())) {
             Allophone existingAllophone = allophoneDao.readByValueIpa(allophone.getValueIpa());
-            if (existingAllophone != null) {
+            if ((existingAllophone != null) && !existingAllophone.getId().equals(allophone.getId())) {
                 result.rejectValue("valueIpa", "NonUnique");
             }
         }
         
         if (StringUtils.isNotBlank(allophone.getValueSampa())) {
             Allophone existingAllophone = allophoneDao.readByValueSampa(allophone.getValueSampa());
-            if (existingAllophone != null) {
+            if ((existingAllophone != null) && !existingAllophone.getId().equals(allophone.getId())) {
                 result.rejectValue("valueSampa", "NonUnique");
             }
         }
@@ -64,10 +66,11 @@ public class AllophoneCreateController {
         if (result.hasErrors()) {
             model.addAttribute("allophone", allophone);
             model.addAttribute("soundTypes", SoundType.values());
-            return "content/sound/create";
+            return "content/sound/edit";
         } else {
             allophone.setTimeLastUpdate(Calendar.getInstance());
-            allophoneDao.create(allophone);
+            allophone.setRevisionNumber(allophone.getRevisionNumber() + 1);
+            allophoneDao.update(allophone);            
             return "redirect:/content/sound/list#" + allophone.getId();
         }
     }
